Move TrainingProgrammeExtension to V2
