+++ src/CommitmentsV2/SFA.DAS.CommitmentsV2.UnitTests/Extensions/TrainingProgramme/WhenDeterminingFundingCap.cs
@@ -1,38 +1,40 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
-using Moq;
 using NUnit.Framework;
-using SFA.DAS.Commitments.Application.Extensions;
-using SFA.DAS.Commitments.Domain.Entities.TrainingProgramme;
+using SFA.DAS.CommitmentsV2.Domain.Entities;
+using SFA.DAS.CommitmentsV2.Extensions;
+using SFA.DAS.CommitmentsV2.Models;
+using SFA.DAS.CommitmentsV2.Types;
 
-namespace SFA.DAS.Commitments.Application.UnitTests.Extensions.ITrainingProgrammeExtensions
+namespace SFA.DAS.CommitmentsV2.UnitTests.Extensions.TrainingProgramme
 {
     [TestFixture]
     public class WhenDeterminingFundingCap
     {
-        private Mock<ITrainingProgramme> _course;
+        private CommitmentsV2.Domain.Entities.TrainingProgramme _course;
 
         [SetUp]
         public void Arrange()
         {
-            _course = new Mock<ITrainingProgramme>();
-            _course.Setup(x => x.EffectiveFrom).Returns(new DateTime(2018, 03, 01));
-            _course.Setup(x => x.EffectiveTo).Returns(new DateTime(2019, 03, 31));
-            _course.Setup(x => x.FundingPeriods).Returns(new List<FundingPeriod>
+            
+            
+            var fundingPeriods = new List<StandardFundingPeriod>
             {
-                new FundingPeriod
+                new StandardFundingPeriod
                 {
                     EffectiveFrom = new DateTime(2018,03,01),
                     EffectiveTo = new DateTime(2018,07,31),
                     FundingCap = 5000
                 },
-                new FundingPeriod
+                new StandardFundingPeriod
                 {
                     EffectiveFrom = new DateTime(2018,08,01),
                     EffectiveTo = null,
                     FundingCap = 2000
                 }
-            });
+            };
+            _course = new CommitmentsV2.Domain.Entities.TrainingProgramme("1", "Test", ProgrammeType.Standard,
+                new DateTime(2018, 03, 01), new DateTime(2019, 03, 31), new List<IFundingPeriod>(fundingPeriods));
         }
 
         [TestCase("2018-05-15", 5000, Description = "Within first funding band")]
@@ -42,7 +44,7 @@ namespace SFA.DAS.Commitments.Application.UnitTests.Extensions.ITrainingProgramm
         public void ThenTheApplicableFundingPeriodIsUsed(DateTime effectiveDate, int expectCap)
         {
             //Act
-            var result = _course.Object.FundingCapOn(effectiveDate);
+            var result = _course.FundingCapOn(effectiveDate);
 
             //Assert
             Assert.AreEqual(expectCap, result);
@@ -57,22 +59,22 @@ namespace SFA.DAS.Commitments.Application.UnitTests.Extensions.ITrainingProgramm
         {
             var courseAndFundingBandStart = new DateTime(2020, 2, 20);
             var courseAndFundingBandEnd = (DateTime?)null;
-
-            _course.Setup(x => x.EffectiveFrom).Returns(courseAndFundingBandStart);
-            _course.Setup(x => x.EffectiveTo).Returns(courseAndFundingBandEnd);
-
-            _course.Setup(x => x.FundingPeriods).Returns(new List<FundingPeriod>
+            var fundingPeriods = new List<StandardFundingPeriod>
             {
-                new FundingPeriod
+                new StandardFundingPeriod
                 {
                     EffectiveFrom = courseAndFundingBandStart,
                     EffectiveTo = courseAndFundingBandEnd,
                     FundingCap = 1
                 }
-            });
+            };
+
+            _course = new CommitmentsV2.Domain.Entities.TrainingProgramme("1", "Test", ProgrammeType.Standard,
+                courseAndFundingBandStart, courseAndFundingBandEnd, new List<IFundingPeriod>(fundingPeriods));
+
 
             //Act
-            var result = _course.Object.FundingCapOn(effectiveDate);
+            var result = _course.FundingCapOn(effectiveDate);
 
             //Assert
             Assert.AreEqual(expectCap, result);
@@ -86,28 +88,26 @@ namespace SFA.DAS.Commitments.Application.UnitTests.Extensions.ITrainingProgramm
         {
             var courseAndFundingBandStart = (DateTime?)null;
             var courseAndFundingBandEnd = (DateTime?)null;
-
-            _course.Setup(x => x.EffectiveFrom).Returns(courseAndFundingBandStart);
-            _course.Setup(x => x.EffectiveTo).Returns(courseAndFundingBandEnd);
-
-            _course.Setup(x => x.FundingPeriods).Returns(new List<FundingPeriod>
+            var fundingPeriods = new List<StandardFundingPeriod>
             {
-                new FundingPeriod
+                new StandardFundingPeriod
                 {
                     EffectiveFrom = courseAndFundingBandStart,
                     EffectiveTo = new DateTime(2018,07,31),
                     FundingCap = 1
                 },
-                new FundingPeriod
+                new StandardFundingPeriod
                 {
                     EffectiveFrom = new DateTime(2018,08,01),
                     EffectiveTo = courseAndFundingBandEnd,
                     FundingCap = 2
                 }
-            });
+            };
+            _course = new CommitmentsV2.Domain.Entities.TrainingProgramme("1", "Test", ProgrammeType.Standard,
+                courseAndFundingBandStart, courseAndFundingBandEnd, new List<IFundingPeriod>(fundingPeriods));
 
             //Act
-            var result = _course.Object.FundingCapOn(effectiveDate);
+            var result = _course.FundingCapOn(effectiveDate);
 
             //Assert
             Assert.AreEqual(expectCap, result);
@@ -124,33 +124,33 @@ namespace SFA.DAS.Commitments.Application.UnitTests.Extensions.ITrainingProgramm
             var courseAndFundingBandStart = new DateTime(2018, 08, 15);
             var courseAndFundingBandEnd = new DateTime(2018, 08, 25);
 
-            _course.Setup(x => x.EffectiveFrom).Returns(courseAndFundingBandStart);
-            _course.Setup(x => x.EffectiveTo).Returns(courseAndFundingBandEnd);
-
-            _course.Setup(x => x.FundingPeriods).Returns(new List<FundingPeriod>
+            
+            var fundingPeriods = new List<StandardFundingPeriod>
             {
-                new FundingPeriod
+                new StandardFundingPeriod
                 {
                     EffectiveFrom = courseAndFundingBandStart,
                     EffectiveTo = new DateTime(2018,08,15),
                     FundingCap = 1
                 },
-                new FundingPeriod
+                new StandardFundingPeriod
                 {
                     EffectiveFrom = new DateTime(2018,08,16),
                     EffectiveTo = new DateTime(2018,08,16),
                     FundingCap = 2
                 },
-                new FundingPeriod
+                new StandardFundingPeriod
                 {
                     EffectiveFrom = new DateTime(2018,08,17),
                     EffectiveTo = courseAndFundingBandEnd,
                     FundingCap = 3
                 }
-            });
+            };
+            _course = new CommitmentsV2.Domain.Entities.TrainingProgramme("1", "Test", ProgrammeType.Standard,
+                courseAndFundingBandStart, courseAndFundingBandEnd, new List<IFundingPeriod>(fundingPeriods));
 
             //Act
-            var result = _course.Object.FundingCapOn(effectiveDate);
+            var result = _course.FundingCapOn(effectiveDate);
 
             //Assert
             Assert.AreEqual(expectCap, result);
@@ -166,33 +166,32 @@ namespace SFA.DAS.Commitments.Application.UnitTests.Extensions.ITrainingProgramm
             var courseAndFundingBandStart = (DateTime?)null;
             var courseAndFundingBandEnd = (DateTime?)null;
 
-            _course.Setup(x => x.EffectiveFrom).Returns(courseAndFundingBandStart);
-            _course.Setup(x => x.EffectiveTo).Returns(courseAndFundingBandEnd);
-
-            _course.Setup(x => x.FundingPeriods).Returns(new List<FundingPeriod>
+            var fundingPeriods = new List<StandardFundingPeriod>
             {
-                new FundingPeriod
+                new StandardFundingPeriod
                 {
                     EffectiveFrom = courseAndFundingBandStart,
                     EffectiveTo = new DateTime(2018,08,15),
                     FundingCap = 1
                 },
-                new FundingPeriod
+                new StandardFundingPeriod
                 {
                     EffectiveFrom = new DateTime(2018,08,16),
                     EffectiveTo = new DateTime(2018,08,20),
                     FundingCap = 2
                 },
-                new FundingPeriod
+                new StandardFundingPeriod
                 {
                     EffectiveFrom = new DateTime(2018,08,21),
                     EffectiveTo = courseAndFundingBandEnd,
                     FundingCap = 3
                 }
-            });
+            };
+            _course = new CommitmentsV2.Domain.Entities.TrainingProgramme("1", "Test", ProgrammeType.Standard,
+                courseAndFundingBandStart, courseAndFundingBandEnd, new List<IFundingPeriod>(fundingPeriods));
 
             //Act
-            var result = _course.Object.FundingCapOn(effectiveDate);
+            var result = _course.FundingCapOn(effectiveDate);
 
             //Assert
             Assert.AreEqual(expectCap, result);
@@ -202,10 +201,10 @@ namespace SFA.DAS.Commitments.Application.UnitTests.Extensions.ITrainingProgramm
         public void IfThereAreNoFundingPeriodsThenCapShouldBeZero()
         {
             //Arrange
-            _course.Setup(x => x.FundingPeriods).Returns(new List<FundingPeriod>());
+            _course.FundingPeriods = new List<TrainingProgrammeFundingPeriod>();
 
             //Act
-            var result = _course.Object.FundingCapOn(new DateTime(2018, 05, 15));
+            var result = _course.FundingCapOn(new DateTime(2018, 05, 15));
 
             //Assert
             Assert.AreEqual(0, result);
@@ -215,7 +214,7 @@ namespace SFA.DAS.Commitments.Application.UnitTests.Extensions.ITrainingProgramm
         public void FundingPeriodsAreEffectiveUntilTheEndOfTheDay()
         {
             //Act
-            var result = _course.Object.FundingCapOn(new DateTime(2018, 7, 31, 23, 59, 59));
+            var result = _course.FundingCapOn(new DateTime(2018, 7, 31, 23, 59, 59));
 
             //Assert
             Assert.AreEqual(5000, result);
