+++ src/CommitmentsV2/SFA.DAS.CommitmentsV2.UnitTests/Extensions/TrainingProgramme/WhenDeterminingWhetherACourseIsActive.cs
@@ -1,10 +1,8 @@
-﻿using System;
-using Moq;
+using System;
 using NUnit.Framework;
-using SFA.DAS.Commitments.Application.Extensions;
-using SFA.DAS.Commitments.Domain.Entities.TrainingProgramme;
+using SFA.DAS.CommitmentsV2.Types;
 
-namespace SFA.DAS.Commitments.Application.UnitTests.Extensions.ITrainingProgrammeExtensions
+namespace SFA.DAS.CommitmentsV2.UnitTests.Extensions.TrainingProgramme
 {
     [TestFixture]
     public class WhenDeterminingWhetherACourseIsActive
@@ -21,12 +19,10 @@ namespace SFA.DAS.Commitments.Application.UnitTests.Extensions.ITrainingProgramm
         public void ThenIfWithinCourseEffectiveRangeThenIsActive(DateTime? courseStart, DateTime? courseEnd, DateTime effectiveDate, bool expectIsActive)
         {
             //Arrange
-            var course = new Mock<ITrainingProgramme>();
-            course.SetupGet(x => x.EffectiveFrom).Returns(courseStart);
-            course.SetupGet(x => x.EffectiveTo).Returns(courseEnd);
-
+            var course = new CommitmentsV2.Domain.Entities.TrainingProgramme("1","test",ProgrammeType.Standard,courseStart,courseEnd);
+            
             //Act
-            var result = course.Object.IsActiveOn(effectiveDate);
+            var result = course.IsActiveOn(effectiveDate);
 
             //Assert
             Assert.AreEqual(expectIsActive, result);
