+++ src/SFA.DAS.Commitments.Application.UnitTests/SFA.DAS.Commitments.Application.UnitTests.csproj
@@ -277,9 +277,6 @@
     <Compile Include="Commands\UpdateProviderPaymentsPriority\WhenValidatingCommand.cs" />
     <Compile Include="Extensions\DataLockStatusExtensions\WhenGettingPriceEpisodeIdentifierDate.cs" />
     <Compile Include="Extensions\DateTimeExtensions\WhenGettingFirstOfMonth.cs" />
-    <Compile Include="Extensions\ITrainingProgrammeExtensions\WhenDeterminingFundingCap.cs" />
-    <Compile Include="Extensions\ITrainingProgrammeExtensions\WhenDeterminingTheStatusOfACourseOnSpecificDate.cs" />
-    <Compile Include="Extensions\ITrainingProgrammeExtensions\WhenDeterminingWhetherACourseIsActive.cs" />
     <Compile Include="Queries\GetActiveApprenticeshipsByUln\WhenGetActiveApprenticeshipsByUln.cs" />
     <Compile Include="Queries\GetActiveApprenticeshipsByUln\WhenValidatingRequest.cs" />
     <Compile Include="Queries\GetCommitmentAgreements\WhenGettingCommitmentAgreements.cs" />
